osugi/material/barcode/requirements.txt
streamlit
pillow
opencv-python
requests
cv2
pyzbar
streamlit-webrtc